api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0